[OV-65] refactor(hub) : requestDto 패키지 수정 및 Update null 체크 제거
- PUT 요청으로 들어오기 때문에 모든 필드가 채워진 상태로 들어오며, dto단에서 validate하기 때문에 null 체크하지 않도록 수정
- XXXRequestDto를 request 패키지로 이동
