+++ hub/src/main/java/com/owlexpress/hub/domain/service/HubServiceImpl.java
@@ -3,8 +3,8 @@ package com.owlexpress.hub.domain.service;
 import com.owlexpress.hub.common.exception.HubException.HubNotFoundException;
 import com.owlexpress.hub.domain.entity.Hub;
 import com.owlexpress.hub.domain.repository.HubRepository;
-import com.owlexpress.hub.presentation.dto.HubCreateRequestDto;
-import com.owlexpress.hub.presentation.dto.HubUpdateRequestDto;
+import com.owlexpress.hub.presentation.dto.request.HubCreateRequestDto;
+import com.owlexpress.hub.presentation.dto.request.HubUpdateRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
